+++ spec/config/config_spec.rb
@@ -60,4 +60,12 @@ describe Cursor::Configuration do
       let(:test_value)    { :test }
     end
   end
+
+  describe 'default_page_by' do
+    it_behaves_like 'a configuration parameter' do
+      let(:config_name)   { :default_page_by }
+      let(:default_value) { :id }
+      let(:test_value)    { :test }
+    end
+  end
 end